+++ dart/packages/fory/lib/src/serializer/struct_serializer.dart
@@ -23,8 +23,8 @@ import 'package:fory/src/meta/field_info.dart';
 import 'package:fory/src/meta/field_type.dart';
 import 'package:fory/src/meta/type_def.dart';
 import 'package:fory/src/resolver/type_resolver.dart';
+import 'package:fory/src/serializer/collection_serializers.dart';
 import 'package:fory/src/serializer/compatible_struct_metadata.dart';
-import 'package:fory/src/serializer/compatible_collection_array_reader.dart';
 import 'package:fory/src/serializer/serializer.dart';
 import 'package:fory/src/serializer/serialization_field_info.dart';
 import 'package:fory/src/serializer/serializer_support.dart';
